refactor(review): 拆分 service_plugin_models.dart (811行) 为三职责模块
- service_field_type.dart：ServiceFieldType 枚举 + key 前缀/组件名映射表 + 推断函数
- service_showhide_rule.dart：ShowHide 规则模型 + 表达式求值器（含日期解析 helper）
- service_plugin_models.dart：保留 DataSource/日期顺序/插件/schema + barrel export，
  外部 import 与测试零改动
